+++ CHANGELOG.md
@@ -3,6 +3,12 @@ Changelog
 
 Summary of all changes made since the first release
 
+0.0.2 (2025-XX-XX)
+------------------
+* ENH: added io submodule with routines to download, load, and write data
+* ENH: added utils submodule with routines to calculate coordinates
+
+
 0.0.1 (2025-08-15)
 ------------------
 * Alpha Release